Add files via upload
While I was in the summer analytics-course I attempted to a hackathon project where we had to create a ML model to predict whether an activity on the website can be considered a Neptune attack or not. I worked on it and got 100 % on the ai planet. This is what actually I'd done.